refactor(dccpeep): extract shared control-flow services
Move shared label/jump parsing, global-label classification, loop-back discovery, bounded label lookup, and conservative internal-label reachability into peep_control_flow.c. Export only is_jump_line in addition to the existing private helper contract.

Validation: strict build clean; six fixtures pass; all 316 raw inputs byte-identical in default, -Os, and -fundocumented-z80 modes.
